Ignore alarm events with the same event time
The counters resolution is smaller than the the event time so random
events occur even if counter_value and alarm_value are equal.

To prevent this it is necessary to ignore events which have an event
time equal to the previous event.

Changing the alarm test_type is necessary or the alarm will stop
working. Therefore do not completely ignore event time, but use it to
decide whether to emit a signal